Changed the weight variable to int

+++ app/src/main/java/com/solusianakbangsa/gameyourfit/AlphaOneActivity.kt
@@ -40,7 +40,7 @@ class AlphaOneActivity : AppCompatActivity(), SensorEventListener {
     private lateinit var levelString: String
     private lateinit var level : JSONObject
 
-    private var weight = 0L
+    private var weight = 0
     private var mAccelerometerLinear: Sensor? = null
     private var exerciseList: MutableList<Signal> = mutableListOf()
     private var counterMax = 0  // Max rep for a certain exercise
@@ -66,7 +66,7 @@ class AlphaOneActivity : AppCompatActivity(), SensorEventListener {
         val toolbar: Toolbar = findViewById(R.id.alphaOneToolbar)
         val sharedPref = PreferenceManager.getDefaultSharedPreferences(this)
         if(sharedPref.contains("weight")){
-            weight = sharedPref.getLong("weight", 57)
+            weight = sharedPref.getInt("weight", 57)
         } else {
             weight = 57
         }
@@ -388,7 +388,7 @@ class AlphaOneActivity : AppCompatActivity(), SensorEventListener {
         repBefore = rep
     }
 
-    private fun countCalorie(time: Long, met: Double, weight: Long) {
+    private fun countCalorie(time: Long, met: Double, weight: Int) {
         // Counts burnt calories for a single exercise session
         val inMinutes = (time.toDouble() / 1000.0) / 60.0
         val caloriesBurned = (met * 3.5 * weight.toDouble() / 200.0) * inMinutes
